@@ -30,85 +30,84 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.10",
-    "@angular/common": "~11.2.10",
-    "@angular/compiler": "~11.2.10",
-    "@angular/core": "~11.2.10",
-    "@angular/forms": "~11.2.10",
-    "@angular/platform-browser": "~11.2.10",
-    "@angular/platform-browser-dynamic": "~11.2.10",
-    "@angular/router": "~11.2.10",
+    "@angular/animations": "~12.0.2",
+    "@angular/common": "~12.0.2",
+    "@angular/compiler": "~12.0.2",
+    "@angular/core": "~12.0.2",
+    "@angular/forms": "~12.0.2",
+    "@angular/platform-browser": "~12.0.2",
+    "@angular/platform-browser-dynamic": "~12.0.2",
+    "@angular/router": "~12.0.2",
     "@ngneat/lib": "^2.0.0",
-    "@ngrx/effects": "^11.1.0",
-    "@ngrx/router-store": "^11.1.0",
-    "@ngrx/store": "^11.1.0",
-    "@reduxjs/toolkit": "^1.5.1",
-    "@rtk-incubator/rtk-query": "^0.3.0",
-    "rxjs": "~6.6.7",
+    "@ngrx/effects": "^12.0.0",
+    "@ngrx/router-store": "^12.0.0",
+    "@ngrx/store": "^12.0.0",
+    "@reduxjs/toolkit": "^1.6.0-beta.0",
+    "rxjs": "~7.1.0",
     "tslib": "^2.2.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^11.1.1",
-    "@angular-devkit/build-angular": "~0.1102.9",
-    "@angular-eslint/builder": "^4.0.0",
-    "@angular-eslint/eslint-plugin": "^4.0.0",
-    "@angular-eslint/eslint-plugin-template": "^4.0.0",
-    "@angular-eslint/schematics": "^4.0.0",
-    "@angular-eslint/template-parser": "^4.0.0",
-    "@angular/cli": "~11.2.9",
-    "@angular/compiler-cli": "~11.2.10",
-    "@commitlint/cli": "^12.1.1",
-    "@commitlint/config-conventional": "^12.1.1",
-    "@ngrx/store-devtools": "^11.1.0",
+    "@angular-builders/jest": "^12.0.0",
+    "@angular-devkit/build-angular": "~12.0.2",
+    "@angular-eslint/builder": "^12.1.0",
+    "@angular-eslint/eslint-plugin": "^12.1.0",
+    "@angular-eslint/eslint-plugin-template": "^12.1.0",
+    "@angular-eslint/schematics": "^12.1.0",
+    "@angular-eslint/template-parser": "^12.1.0",
+    "@angular/cli": "~12.0.2",
+    "@angular/compiler-cli": "~12.0.2",
+    "@commitlint/cli": "^12.1.4",
+    "@commitlint/config-conventional": "^12.1.4",
+    "@ngrx/store-devtools": "^12.0.0",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "@tailwindcss/forms": "^0.3.2",
-    "@testing-library/angular": "^10.5.0",
-    "@testing-library/jest-dom": "^5.11.10",
-    "@testing-library/user-event": "^13.1.5",
-    "@types/jest": "^26.0.22",
-    "@types/node": "^12.20.7",
+    "@testing-library/angular": "^10.8.1",
+    "@testing-library/jest-dom": "^5.12.0",
+    "@testing-library/user-event": "^13.1.9",
+    "@types/jest": "^26.0.23",
+    "@types/node": "^12.20.13",
     "@types/node-fetch": "^2.5.10",
-    "@typescript-eslint/eslint-plugin": "^4.22.0",
-    "@typescript-eslint/parser": "^4.22.0",
+    "@typescript-eslint/eslint-plugin": "^4.25.0",
+    "@typescript-eslint/parser": "^4.25.0",
     "all-contributors-cli": "^6.20.0",
     "angular-cli-ghpages": "^1.0.0-rc.1",
     "cpx": "^1.5.0",
     "cross-env": "^7.0.3",
-    "eslint": "^7.24.0",
-    "eslint-config-prettier": "^8.2.0",
-    "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-jest": "^24.3.5",
-    "eslint-plugin-jest-dom": "^3.8.0",
-    "eslint-plugin-jsdoc": "32.3.0",
-    "eslint-plugin-ngrx": "^1.17.4",
+    "eslint": "^7.27.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-jest": "^24.3.6",
+    "eslint-plugin-jest-dom": "^3.9.0",
+    "eslint-plugin-jsdoc": "35.1.0",
+    "eslint-plugin-ngrx": "^1.24.0",
     "eslint-plugin-prefer-arrow": "1.2.3",
     "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-rxjs": "^3.1.5",
-    "eslint-plugin-testing-library": "^4.1.0",
+    "eslint-plugin-rxjs": "^3.3.3",
+    "eslint-plugin-testing-library": "^4.6.0",
     "git-cz": "^4.7.6",
     "husky": "^6.0.0",
     "jest": "^26.6.3",
     "jest-preset-angular": "^8.4.0",
     "jsonc-parser": "^3.0.0",
-    "lint-staged": "^10.5.4",
-    "msw": "^0.28.1",
-    "ng-packagr": "^11.2.4",
+    "lint-staged": "^11.0.0",
+    "msw": "^0.29.0",
+    "ng-packagr": "^12.0.2",
     "node-fetch": "^2.6.1",
-    "postcss": "^8.2.10",
-    "prettier": "^2.2.1",
+    "postcss": "^8.3.0",
+    "prettier": "^2.3.0",
     "protractor": "~7.0.0",
-    "semantic-release": "^17.4.2",
-    "stylelint": "^13.12.0",
+    "semantic-release": "^17.4.3",
+    "stylelint": "^13.13.1",
     "stylelint-config-prettier": "^8.0.2",
     "stylelint-config-recommended-scss": "^4.2.0",
-    "stylelint-config-standard": "^21.0.0",
+    "stylelint-config-standard": "^22.0.0",
     "stylelint-scss": "^3.19.0",
-    "tailwindcss": "^2.1.1",
-    "ts-jest": "^26.5.5",
+    "tailwindcss": "^2.1.2",
+    "ts-jest": "^26.5.6",
     "ts-node": "~8.3.0",
-    "typescript": "~4.1.5"
+    "typescript": "~4.2.4"
   },
   "config": {
     "commitizen": {
