@@ -27,18 +27,20 @@ const importModuleSet: {
   },
 ];
 
-export const ngAdd = (options: Schema): Rule => (tree: Tree) => {
-  const workspaceConfig = tree.read('/angular.json');
-  if (!workspaceConfig) {
-    throw new SchematicsException('Could not find Angular workspace configuration');
-  }
-  return chain([
-    addPackageJsonDependencies(),
-    installPackageJsonDependencies(),
-    injectImports(options),
-    addModuleToImports(options),
-  ]);
-};
+export const ngAdd =
+  (options: Schema): Rule =>
+  (tree: Tree) => {
+    const workspaceConfig = tree.read('/angular.json');
+    if (!workspaceConfig) {
+      throw new SchematicsException('Could not find Angular workspace configuration');
+    }
+    return chain([
+      addPackageJsonDependencies(),
+      installPackageJsonDependencies(),
+      injectImports(options),
+      addModuleToImports(options),
+    ]);
+  };
 
 const addPackageJsonDependencies = (): Rule => (host: Tree, context: SchematicContext) => {
   const dependencies: { name: string; version: string }[] = [];
@@ -58,77 +60,81 @@ const installPackageJsonDependencies = (): Rule => (host: Tree, context: Schemat
   return host;
 };
 
-const injectImports = (options: Schema): Rule => (host: Tree, context: SchematicContext) => {
-  if (!options.skipImport) {
-    const workspace = getWorkspace(host) as any;
-    const project = getProjectFromWorkspace(
-      workspace,
-      options.project ? options.project : Object.keys(workspace.projects)[0],
-    );
+const injectImports =
+  (options: Schema): Rule =>
+  (host: Tree, context: SchematicContext) => {
+    if (!options.skipImport) {
+      const workspace = getWorkspace(host) as any;
+      const project = getProjectFromWorkspace(
+        workspace,
+        options.project ? options.project : Object.keys(workspace.projects)[0],
+      );
 
-    if (!project || project.projectType !== 'application') {
-      throw new SchematicsException(`A client project type of "application" is required.`);
-    }
+      if (!project || project.projectType !== 'application') {
+        throw new SchematicsException(`A client project type of "application" is required.`);
+      }
 
-    if (
-      !project.architect ||
-      !project.architect.build ||
-      !project.architect.build.options ||
-      !project.architect.build.options.main
-    ) {
-      throw targetBuildNotFoundError();
-    }
+      if (
+        !project.architect ||
+        !project.architect.build ||
+        !project.architect.build.options ||
+        !project.architect.build.options.main
+      ) {
+        throw targetBuildNotFoundError();
+      }
 
-    const modulePath = getAppModulePath(host, project.architect.build.options.main);
-    const moduleSource = getSourceFile(host, modulePath);
+      const modulePath = getAppModulePath(host, project.architect.build.options.main);
+      const moduleSource = getSourceFile(host, modulePath);
 
-    importModuleSet.forEach((item) => {
-      if (isImported(moduleSource, item.moduleName, item.importPath)) {
-        context.logger.warn(`Could not import "${item.moduleName}" because it's already imported.`);
-      } else {
-        const change = insertImport(moduleSource, modulePath, item.moduleName, item.importPath);
+      importModuleSet.forEach((item) => {
+        if (isImported(moduleSource, item.moduleName, item.importPath)) {
+          context.logger.warn(`Could not import "${item.moduleName}" because it's already imported.`);
+        } else {
+          const change = insertImport(moduleSource, modulePath, item.moduleName, item.importPath);
 
-        if (change) {
-          const recorder = host.beginUpdate(modulePath);
-          recorder.insertLeft((change as InsertChange).pos, (change as InsertChange).toAdd);
-          host.commitUpdate(recorder);
-          context.logger.log('info', '✅ Written import statement for "' + item.moduleName + '"');
+          if (change) {
+            const recorder = host.beginUpdate(modulePath);
+            recorder.insertLeft((change as InsertChange).pos, (change as InsertChange).toAdd);
+            host.commitUpdate(recorder);
+            context.logger.log('info', '✅ Written import statement for "' + item.moduleName + '"');
+          }
         }
+      });
+      return host;
+    }
+  };
+
+const addModuleToImports =
+  (options: Schema): Rule =>
+  (host: Tree, context: SchematicContext) => {
+    if (!options.skipImport) {
+      const workspace = getWorkspace(host) as any;
+      const project = getProjectFromWorkspace(
+        workspace,
+        options.project ? options.project : Object.keys(workspace.projects)[0],
+      );
+
+      if (!project || project.projectType !== 'application') {
+        throw new SchematicsException(`A client project type of "application" is required.`);
       }
-    });
+      if (!project.architect) {
+        throw new SchematicsException(`Architect options not present for project.`);
+      }
+      if (!project.architect.build) {
+        throw new SchematicsException(`Architect:Build options not present for project.`);
+      }
+
+      const modulePath = getAppModulePath(host, project.architect.build.options.main);
+
+      importModuleSet.forEach((item) => {
+        if (hasNgModuleImport(host, modulePath, item.moduleName)) {
+          context.logger.warn(`Could not set up "${item.moduleName}" in "imports[]" because it's already imported.`);
+        } else {
+          addModuleImportToRootModule(host, item.importModuleStatement, null as any, project);
+          context.logger.log('info', '✅ Imported "' + item.moduleName + '" in imports');
+        }
+      });
+    }
+
     return host;
-  }
-};
-
-const addModuleToImports = (options: Schema): Rule => (host: Tree, context: SchematicContext) => {
-  if (!options.skipImport) {
-    const workspace = getWorkspace(host) as any;
-    const project = getProjectFromWorkspace(
-      workspace,
-      options.project ? options.project : Object.keys(workspace.projects)[0],
-    );
-
-    if (!project || project.projectType !== 'application') {
-      throw new SchematicsException(`A client project type of "application" is required.`);
-    }
-    if (!project.architect) {
-      throw new SchematicsException(`Architect options not present for project.`);
-    }
-    if (!project.architect.build) {
-      throw new SchematicsException(`Architect:Build options not present for project.`);
-    }
-
-    const modulePath = getAppModulePath(host, project.architect.build.options.main);
-
-    importModuleSet.forEach((item) => {
-      if (hasNgModuleImport(host, modulePath, item.moduleName)) {
-        context.logger.warn(`Could not set up "${item.moduleName}" in "imports[]" because it's already imported.`);
-      } else {
-        addModuleImportToRootModule(host, item.importModuleStatement, null as any, project);
-        context.logger.log('info', '✅ Imported "' + item.moduleName + '" in imports');
-      }
-    });
-  }
-
-  return host;
-};
+  };
