@@ -1,8 +1,10 @@
-import { EndpointDefinitions, MutationDefinition, QueryDefinition } from '@rtk-incubator/rtk-query';
-import { DefinitionType } from '@rtk-incubator/rtk-query/dist/esm/ts/endpointDefinitions';
-import { UseLazyQuery, UseMutation, UseQuery } from './hooks-types';
+import type { EndpointDefinitions, MutationDefinition, QueryDefinition } from '@reduxjs/toolkit/query';
+import type { DefinitionType } from '@reduxjs/toolkit/dist/query/endpointDefinitions';
+import type { UseLazyQuery, UseMutation, UseQuery } from './hooks-types';
 
-export type TS41Hooks<Definitions extends EndpointDefinitions> = keyof Definitions extends infer Keys
+// For TS 4.1 and later, we can use string literal types to define
+// the exact names of each endpoint's exported hooks
+export type HooksWithUniqueNames<Definitions extends EndpointDefinitions> = keyof Definitions extends infer Keys
   ? Keys extends string
     ? Definitions[Keys] extends { type: DefinitionType.query }
       ? {
